+++ Shared/timegm.cpp
@@ -48,7 +48,7 @@ int TimeGM::leap_days(int y1, int y2) {
   return (y2 / 4 - y1 / 4) - (y2 / 100 - y1 / 100) + (y2 / 400 - y1 / 400);
 }
 
-time_t TimeGM::parse_fractional_seconds(uint sfrac,
+time_t TimeGM::parse_fractional_seconds(uint64_t sfrac,
                                         const int ntotal,
                                         const SQLTypeInfo& ti) {
   int dimen = ti.get_dimension();
